web/requirements.txt
Django==1.8.3
PAM==0.4.2
Pillow==2.7.0
Twisted-Core==14.0.2
Twisted-Web==14.0.2
adium-theme-ubuntu==0.3.4
argparse==1.2.1
beautifulsoup4==4.3.2
characteristic==0.1.0
chardet==2.3.0
colorama==0.3.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.1.3
djangorestframework-gis==0.9.3
duplicity==0.7.1
gunicorn==19.3.0
html5lib==0.999
httplib2==0.9
idna==0.9
lockfile==0.8
lxml==3.4.4
ndg-httpsclient==0.3.2
numpy==1.8.2
oauthlib==0.6.1
pexpect==3.2
piston-mini-client==0.7.5
psycopg2==2.6.1
pyOpenSSL==0.13.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pycups==1.9.72
pygobject==3.14.0
pyinotify==0.9.5
pykml==0.1.0
pyserial==2.6
python-debian==0.1.22
pyxdg==0.25
requests==2.4.3
service-identity==1.0.0
six==1.9.0
software-center-aptd-plugins==0.0.0
static3==0.6.1
unity-lens-photos==1.0
urllib3==1.9.1
virtualenv==13.1.0
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.7.2
zope.interface==4.1.1
